+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Added a new function "addCustomNetworkRequestHeaders: customHeaderValues" for providing or overriding custom headers after init.
 * Default request method is now set to "POST"
 * Updated user properties caching mechanism according to sessions.
+* Added a new function "recordMetrics: metricsOverride" to send a device metrics request.
+* Added a new Consent option "metrics" for controlling "recordMetrics" method. (This has no effect on Session metrics.)
 * Improved content error handling and display mechanics in Android.
 * Non window contexts also now uses POST requests by default in Web.